Separated views in routers directory - #4
